requirements.txt
Flask==3.0.3
gunicorn==23.0.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
plotly==5.24.1
requests==2.32.3
yfinance==0.2.43
tenacity==9.0.0
python-dateutil==2.8.2